+++ com.unity.visualeffectgraph/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [6.7.0-preview] - 2019-XX-XX
 ### Added
 - Distortion Outputs (Quad / Mesh)
-- Texture Color mode for unlit outputs (Textured/Gradient Mapped)
+- Color mapping mode for unlit outputs (Textured/Gradient Mapped)
 
 ### Fixed
 - Fix shader compilation error with debug views